+++ package.json
@@ -228,7 +228,7 @@
     "typescript": "^2.1.5",
     "uglify-es": "3.3.9",
     "uglifyjs-webpack-plugin": "1.2.2",
-    "vsce": "^1.37.5",
+    "vsce": "^1.37.6",
     "vscode": "^1.1.11",
     "webpack": "^4.1.0",
     "webpack-cli": "^2.0.10",
